move current tests to integration folder, add unit test for expiration calculation
